Standup Writer/requirements.txt
fastapi
uvicorn
openai
python-dotenv
jinja2
python-multipart